requirements.txt
Flask>=0.10.1
leancloud-sdk>=1.3.7
BeautifulSoup4
requests
Image
html5lib
wechat-sdk